+++ lib/eve_online/esi/models/allies.rb
@@ -4,6 +4,9 @@ module EveOnline
   module ESI
     module Models
       class Allies < Base
+        def allies
+
+        end
       end
     end
   end